pyproject.toml
[project]
name = "research-digest"
version = "0.1.0"
description = "A research funnel engine for arXiv papers"
readme = "README.md"
authors = [
    { name = "davhofer", email = "davhofer@ethz.ch" }
]
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "groq>=1.0.0",
    "jinja2>=3.1.6",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "typer>=0.21.1",
]

[project.scripts]
research-digest = "research_digest.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "research-digest",
]

[tool.uv.sources]
research-digest = { workspace = true }
